% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{doe_plot}
\alias{doe_plot}
\title{DoE Plot}
\usage{
doe_plot(ndt_full_res, doe_type = "1")
}
\arguments{
\item{ndt_full_res}{The output of running run_full_ndt()}

\item{doe_type}{Whether to include (doe_type="1") or ignore (doe_type="2") the
contribution of dark uncertainty.}
}
\description{
Outputs DoE plot.
}
