% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_doe_table}
\alias{get_doe_table}
\title{DoE Table}
\usage{
get_doe_table(ndt_full_res, doe_type = "1")
}
\arguments{
\item{ndt_full_res}{The output of running run_full_ndt()}

\item{doe_type}{Whether to include (doe_type="1") or ignore (doe_type="2") the
contribution of dark uncertainty.}
}
\description{
Returns the DoE Table
}
