% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_KCplot}
\alias{get_KCplot}
\title{Create KCRV and Lab Uncertainty Plot}
\usage{
get_KCplot(ndt_full_res)
}
\arguments{
\item{ndt_full_res}{The output of running run_full_ndt()}
}
\description{
Creates a plot of the KCRV estimate and uncertainty, along with labs
and their uncertainties.
}
