% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_MCMC_diagnostics}
\alias{get_MCMC_diagnostics}
\title{MCMC Diagnostics}
\usage{
get_MCMC_diagnostics(ndt_full_res)
}
\arguments{
\item{ndt_full_res}{The output of running run_full_ndt()}
}
\description{
If applicable, prints n_eff and Rhat from the MCMC sampler.
}
