% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{run_full_ndt}
\alias{run_full_ndt}
\title{Run the full NDT procedure}
\usage{
run_full_ndt(
  dataset,
  exclude,
  procedure = "Recommended",
  num_bootstrap = 1000,
  seed = 123,
  n_iter = 50000,
  burn_in = 25000,
  thin = 10
)
}
\arguments{
\item{dataset}{An R data.frame with columns titled "Laboratory", "MeasuredValues", "StdUnc", and "DegreesOfFreedom"}

\item{exclude}{A boolean vector of which labs to exclude from the KCRV computation.}

\item{procedure}{Which procedure to run. E.g., "Recommended", "AWA","WM", "HGG", etc.}

\item{num_bootstrap}{The number of bootstrap iterations, for the AWA and WM procedures.}

\item{seed}{The random seed used for all simulations and mcmc.}

\item{n_iter}{The total number of MCMC iterations.}

\item{burn_in}{The number of burn-in MCMC iterations.}

\item{thin}{The thinning rate for the MCMC iterations.}
}
\value{
A list of objects containing results pertaining to the analysis.
}
\description{
Runs the full NIST Decision Tree Procedure
and outputs the results as a list. The output can be inspected
directly and also used to generate tables and plots, e.g., get_KCplot()
}
