% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{summary_table}
\alias{summary_table}
\title{Lab Uncertainties Table}
\usage{
summary_table(
  ndt_full_res = NULL,
  ndt_res = NULL,
  vars_in = NULL,
  doe_table = NULL
)
}
\arguments{
\item{ndt_full_res}{The output of running run_full_ndt()}

\item{ndt_res}{Used only for the NDT Shiny app. Leave as NULL for use from
within R.}

\item{vars_in}{Used only for the NDT Shiny app. Leave as NULL for use from
within R.}

\item{doe_table}{Used only for the NDT Shiny app. Leave as NULL for use from
within R.}
}
\description{
Returns a dataframe of uncertainties associated with each lab,
for both the lab means and DoE. Uncertainties are returned for
cases where dark uncertainty is both included and ignored.
}
